refactor: 단일 LottoView 클래스를 4개의 하위 뷰로 분할
